requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.2
python-socketio==5.7.2
python-engineio==4.3.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.4
pandas==1.5.3
numpy==1.24.2
plotly==5.13.1
gunicorn==20.1.0
pytest==7.2.2
python-dotenv==1.0.0
click==8.1.3 